% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateFood.R
\name{calculateFood}
\alias{calculateFood}
\title{Calculate Food}
\usage{
calculateFood(food, elas, gdp_pc, functionalForm, pop, referenceYear, dt)
}
\arguments{
\item{food}{The food consumption at time t.}

\item{elas}{The elasticity of the commodity.  This will vary by country and 
commodity.}

\item{gdp_pc}{Per person GDP (gross domestic product) at time t.}

\item{functionalForm}{Currently one of 0, 1, 2, or 3.  Specifies if a log-log,
semi-log, or inverse-log food demand model should be used.}

\item{pop}{Population at time t.}

\item{referenceYear}{Year of reference to compute the food estimates.}

\item{dt}{Table in data.table format}

\item{trend_factor}{Changes due to income and other factors.}
}
\value{
A numeric vector of the estimated food consumption.
}
\description{
Calculate Food
}

