% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/center.R
\name{center}
\alias{center}
\title{Center
These four different functions provide estimates for the food consumption 
values in year t+1 given the consumption in year t, changes in income, and 
the elasticity of the particular commodity.}
\usage{
center(type = c("0", "1", "2", "3", "NA"), ...)
}
\arguments{
\item{type}{A character value.}
}
\description{
Center
These four different functions provide estimates for the food consumption 
values in year t+1 given the consumption in year t, changes in income, and 
the elasticity of the particular commodity.
}

