% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commodity2FunctionalForm.R
\name{commodity2FunctionalForm}
\alias{commodity2FunctionalForm}
\title{Commodity to Functional Form}
\usage{
commodity2FunctionalForm(commodityCode)
}
\arguments{
\item{commodityCode}{A character vector of commodity codes.}
}
\value{
A list with two objects: a vector of the corresponding food demand 
  codes and a vector of the corresponding food commodity codes.
}
\description{
This function takes a commodity code and maps the commodity code to a "food 
demand model" code and a "food commodity model" code.  These are types of
models that were originally implemented by Josef Schmidhuber in SAS, and we
are re-using the coefficients in R in the absence of a better food model.
}

