% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeFoodForwardBackward.R
\name{computeFoodForwardBackward}
\alias{computeFoodForwardBackward}
\title{Calculate Food Forward and Backward}
\usage{
computeFoodForwardBackward(food, pop, elas, gdp, netTrade, functionalForm,
  timePointYears, protected, type, referenceYear)
}
\arguments{
\item{food}{The food consumption at time t.}

\item{pop}{Population at time t.}

\item{elas}{The elasticity of the commodity.  This will vary by country and 
commodity.}

\item{netTrade}{The Net Trade (imports - exports) at time t.}

\item{functionalForm}{Currently one of 0, 1, 2, or 3.  Specifies if a log-log,
semi-log, or inverse-log food demand model should be used.}

\item{timePointYears}{A numeric vector with the years.}

\item{protected}{A logical value indicating whether the value is proetcted or not.}

\item{type}{A character string indicating the levels of the food classification of the commodity.}

\item{referenceYear}{Year of reference to start to compute the food estimates.}

\item{gdp_pc}{Per person GDP (gross domestic product) at time t.}
}
\value{
A numeric vector of the estimated food consumption.
}
\description{
The function compute the estimates for food consumption in the time t based 
on the food consumption in the t-1.
}

