% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FoodModule.R
\name{FoodModule}
\alias{FoodModule}
\title{Food Module}
\usage{
FoodModule(data)
}
\arguments{
\item{data}{A data.table object containing the food data.}
}
\value{

}
\description{
This function takes a dataset with the appropriate variables (food,
population, GDP, foodFunction, elasticity, and foodPerCapita) and
estimates food values (and standard errors) for the next year
}

